+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 0.66.0
 ------
 - Style changes
+    - Updated `--color base16` (alias: `16`) theme so that it works better with both dark and light themes.
     - Narrowed the gutter column by using the left-half block character (`▌`).
     - Removed background colors from markers.
 - Added `--gutter CHAR` option for customizing the gutter column. Some examples using [box-drawing characters](https://en.wikipedia.org/wiki/Box-drawing_characters):